refactor: Upgrade V2 Logger to V3
