+++ gradle/libs.versions.toml
@@ -46,7 +46,7 @@ firebaseBom = "34.11.0"
 crashlytics = "3.0.3"
 classgraph = "4.8.179"
 govuk-ui = "16.10.0"
-govuk-logging = "0.40.2"
+govuk-logging = "0.40.4"
 compose-runtime = "1.9.4"
 auth = "3.3.0"
 wallet = "v2.292.1"
@@ -140,6 +140,7 @@ authentication = { group = "uk.gov.android.authentication", name = "app", versio
 localauth = { group = "uk.gov.android.authentication", name = "localauth", version.ref = "auth" }
 secure-store = { group = "uk.gov.securestore", name = "app", version.ref = "secureStore" }
 network = { group = "uk.gov.android", name = "network", version.ref = "network" }
+logging-api = { group = "uk.gov.logging", name = "logging-api", version.ref = "govuk-logging" }
 logging-impl = { group = "uk.gov.logging", name = "logging-impl", version.ref = "govuk-logging" }
 logging-test = { group = "uk.gov.logging", name = "logging-testdouble", version.ref = "govuk-logging" }
